ReplicaSet: Use apps/v1 RS in kube-controller-manager.
